Move Fleet resource package to managedintegration (OpenSpec task 3)
Rename the experimental resource registration to elasticstack_fleet_managed_integration while preserving git history for the former agentlesspolicy package.
