+++ internal/fleet/managedintegration/capabilities.go
@@ -15,7 +15,7 @@
 // specific language governing permissions and limitations
 // under the License.
 
-package agentlesspolicy
+package managedintegration
 
 import (
 	"context"
