+++ internal/fleet/managedintegration/enabled_convergence_test.go
@@ -15,7 +15,7 @@
 // specific language governing permissions and limitations
 // under the License.
 
-package agentlesspolicy
+package managedintegration
 
 import (
 	"context"
