+++ internal/fleet/managedintegration/kbapi_roundtrip_test.go
@@ -15,7 +15,7 @@
 // specific language governing permissions and limitations
 // under the License.
 
-package agentlesspolicy
+package managedintegration
 
 import (
 	"encoding/json"
